+++ scripts/macros/2024/spells/circleOfPower.js
@@ -1,0 +1,5 @@
+import {circleOfPower as circleOfPowerLegacy} from '../../2014/spells/circleOfPower.js';
+export let circleOfPower = {
+    ...circleOfPowerLegacy,
+    rules: 'modern'
+};